+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintExtension.kt
@@ -1,5 +1,6 @@
 package org.jlleitschuh.gradle.ktlint
 
+@Deprecated("Moved to org.jlleitschuh.gradle.ktlint.reporter package")
 typealias ReporterType = org.jlleitschuh.gradle.ktlint.reporter.ReporterType
 
 /**
@@ -41,18 +42,18 @@ open class KtlintExtension {
      *
      * **Note** for Gradle scripts: for now all values should be uppercase due to bug in Gradle.
      *
-     * Default is empty.
+     * Default is plain.
      */
-    var reporters: Array<ReporterType> = emptyArray()
+    var reporters: Array<ReporterType> = arrayOf(ReporterType.PLAIN)
 
     /**
      * Report output format.
      *
      * Available values: plain, plain_group_by_file, checkstyle, json.
      *
-     * Default is plain.
+     * Default is none, please, mirgrate to [reporters].
      */
     @Deprecated(message = "Ktlint introduced multi output support since 0.11.1 version",
             replaceWith = ReplaceWith("reporters = arrayOf()", "reporter"))
-    var reporter: ReporterType? = ReporterType.PLAIN
+    var reporter: ReporterType? = null
 }