+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - Output to console (#38)
  - Support multiple reporters for ktlint >0.10.x (#38)
 ### Changed
- - Set default ktltint version to `0.13.0`
+ - Set default ktltint version to `0.15.0`
 
 ## [3.0.0] - 2017-12-25
 ## Changed